fix(fundraiser filter/search): fix fundraiser filter and search
